requirements.txt
alembic>=1.0.9,<1.1
blessings>=1.7,<1.8
blinker>=1.4,<1.5
flask>=1.0.2,<1.1
flask-heroku>=0.1.9,<0.2.0
flask-sslify>=0.1.5,<0.2.0
flask-sqlalchemy>=2.3.2,<2.4
flask-redis>=0.3,<0.4
flask-restful>=0.3.7,<0.4.0
oauth2client>=4.1.3,<4.2
psycopg2-binary>=2.8.2,<2.9
raven==6.10.0
redis>=3.2.1,<3.3
requests>=2.21,<2.22
rq>=1.0,<1.1
werkzeug>=0.15.2,<0.16
